made the application threadsafe and used an efficient datastructure

+++ src/main/java/word/counter/service/WordCounterService.java
@@ -4,7 +4,8 @@ import word.counter.exception.IllegalCharacterException;
 
 public interface WordCounterService {
 
+
     public void addWords(String word) throws IllegalCharacterException;
 
-    public long countSimilarMeaningWords(String word);
+    public int countSimilarMeaningWords(String word);
 }
